+++ CPU/test.cpp
@@ -1,4 +1,4 @@
 #include <vector>
 #include "test.h"
 
-test::test(std::vector<double> const& features, double anwser) : features(features), anwser(anwser) {}
+test::test(std::vector<float> const& features, float anwser) : features(features), anwser(anwser) {}
